[Bronze V] Title: 이상한 기호, Time: 204 ms, Memory: 17696 KB -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -5,8 +5,8 @@ public class Main {
         Scanner sc = new Scanner(System.in);
         String str = sc.nextLine();
         String[] result = str.split(" ");
-        int A = Integer.parseInt(result[0]);
-        int B = Integer.parseInt(result[1]);
+        Long A = Long.parseLong(result[0]);
+        Long B = Long.parseLong(result[1]);
         System.out.println((A+B)*(A-B));
     }
 }
